# Add matmul optimization for the case A.ndim <= 2 && B.ndim >= 3