… test/common_methods_invocations.py