This addresses #18862.