@skrah 2019-05-13
This PR implements more or less @colesbury's suggestion **C** from #18862.  Suggestion **C** ends up calling the existing optimization for `A.ndim >= 3 && B.ndim <= 2` after transposing the inner dimensions of the arguments and swapping them.

***Memory usage***

Memory usage as reported by `valgrind --tool=massif`.

**Test case**

```
x = torch.randn(4096, 4096)
y = torch.randn(192, 4096, 1)
z = torch.matmul(x, y)
```

**Before**

```
--------------------------------------------------------------------------------
  n        time(i)         total(B)   useful-heap(B) extra-heap(B)    stacks(B)
--------------------------------------------------------------------------------
 51 18,481,148,839   12,980,531,144   12,978,674,895     1,856,249            0
 52 18,557,183,694   12,980,531,144   12,978,674,895     1,856,249            0
 53 18,633,218,549   12,980,531,144   12,978,674,895     1,856,249            0
 54 18,688,264,583       17,333,504       15,568,202     1,765,302            0
```

**After**

```
--------------------------------------------------------------------------------
  n        time(i)         total(B)   useful-heap(B) extra-heap(B)    stacks(B)
--------------------------------------------------------------------------------
 76  5,124,473,874       96,953,336       95,088,436     1,864,900            0
 77  5,139,417,005       95,591,328       93,726,580     1,864,748            0
 78  5,155,408,421       22,011,848       20,148,194     1,863,654            0
 79  5,171,971,669       21,145,472       19,304,096     1,841,376            0
 80  5,180,817,596       17,866,168       16,086,322     1,779,846            0
 81  5,189,662,538        7,099,760        6,442,537       657,223            0
```

***Timings****

**Test case**

```
x = torch.randn(4096, 4096)
y = torch.randn(192, 4096, 1)

with torch.autograd.profiler.profile(use_cuda=True) as prof:
    z = torch.matmul(x, y)

print(prof)
```

**Before**

```
Self CPU time total: 7.430s
CUDA time total: 7.430s
```

**After**

```
Self CPU time total: 445.886ms
CUDA time total: 445.883ms
```