@skrah 2019-05-13
For small matrices the speed is about the same:

```
x = torch.randn(64)
y = torch.randn(2, 64, 1)

with torch.autograd.profiler.profile(use_cuda=True) as prof:
    for i in range(10000):
        z = torch.matmul(x, y)
```

**Before**

```
Self CPU time total: 1.076s
CUDA time total: 1.081s
```

**After**

```
Self CPU time total: 939.033ms
CUDA time total: 940.369ms
```