@skrah 2019-05-14
The PR is conservative in not reusing the `out_opt` argument in the recursive call, but I don't see a big difference in the timings (both the old and the new code have the same timings with or without and explicit `out` arg).