@skrah 2019-05-15
`test_torch.py` already has tests here (they fail if an error is introduced in the new code):

https://github.com/pytorch/pytorch/blob/8e26759f14c9e4e310e05058fcb687ddc075825c/test/test_torch.py#L4137

I've added some basic `test_autograd.py` tests.


Also I've tested the code with a couple of throwaway scripts like this one, but they may be too long for the unit tests:

```
import torch
import numpy as np

# 2d, 3d
for N in range(1, 20):
  for M in range(1, 20):
    for P in range(1, 20):
      for O in range(1, 20):
        x = torch.arange(N*M).reshape(N, M)
        y = torch.arange(O*M*P).reshape(O, M, P)
        expected = torch.bmm(x.unsqueeze(0).expand(O, N, M), y)
        z = torch.matmul(x, y)
        if not torch.equal(z, expected):
          raise RuntimeError("different results: %s %s %s %s" % (N, M, P, O))
        # Check contiguity flags via numpy.
        ex = np.array(expected, copy=False)
        zz = np.array(z, copy=False)
        if ex.flags != zz.flags or ex[0].flags != zz[0].flags:
          raise RuntimeError("different flags: %s %s %s %s" % (N, M, P, O))

# 1d, 3d
N = 1
for M in range(1, 20):
  for P in range(1, 20):
    for O in range(1, 20):
      x = torch.arange(M)
      y = torch.arange(O*M*P).reshape(O, M, P)
      expected = torch.bmm(x.expand(O, N, M), y).reshape(O, P)
      z = torch.matmul(x, y)
      if not torch.equal(z, expected):
        raise RuntimeError("different results: %s %s %s %s" % (N, M, P, O))
      # Check contiguity flags via numpy.
      ex = np.array(expected, copy=False)
      zz = np.array(z, copy=False)
      if ex.flags != zz.flags or  ex[0].flags != zz[0].flags:
        raise RuntimeError("different flags: %s %s %s %s" % (N, M, P, O))
```