@skrah 2019-05-15
@colesbury Thanks for the comments, I think they have all been addressed.