@ezyang 2019-05-15
> Also I've tested the code with a couple of throwaway scripts like this one, but they may be too long for the unit tests

That's what the `@slowTest` decorator is for :)