@skrah 2019-05-15
> That's what the `@slowTest` decorator is for :)

OK thanks, actually I can add that test with `range(1, 10)`, then it just takes 1s (4s in a debug build).

 Apparently `test_matmul_4d_4d` was banned at some point, it is in a list `THESE_TAKE_WAY_TOO_LONG` and I can no longer find the test itself. :)